// src/shared/api/base/generated-types.ag.ts
export enum PSupportedLocaleIds {
	de_DE = 'de-DE',
	de_AT = 'de-AT',
	de_CH = 'de-CH',
	en_NL = 'en-NL',
	en_BE = 'en-BE',
	en_GB = 'en-GB',
	en_CZ = 'en-CZ',
	en_SE = 'en-SE',
	de_LU = 'de-LU',
	en = 'en',
}


/**
 * All languages that are available in our App (first two chars from LOCALE_ID).
 */
export enum PSupportedLanguageCodes {
	de = 'de',
	en = 'en',
}

/**
 * All country-codes that are available in our App (last two chars from LOCALE_ID).
 */
export enum PSupportedCountryCodes {
	DE = 'DE',
	AT = 'AT',
	CH = 'CH',
	NL = 'NL',
	BE = 'BE',
	GB = 'GB',
	CZ = 'CZ',
	SE = 'SE',
	LU = 'LU',
}
// ...
export enum PSupportedCurrencyCodes {
	EUR = 'EUR',
	CHF = 'CHF',
	GBP = 'GBP',
	CZK = 'CZK',
	SEK = 'SEK',
}

/**
 * All time-zones that are available in our App.
 */
export enum PSupportedTimeZones {
	EUROPE_BERLIN = 'Europe/Berlin',
	ETC_GMT = 'Etc/GMT',
}